Split .update to .update_only & .to_euler_angles
This PR builds on top of #7 and splits `.update` to `.update_only` &
`.to_euler_angles`.  `.update_only` is similar to `.update`, but the
current estimates (roll, pitch, yaw) are not returned from this method
and instead have to be explicitly queried with `.to_euler_angles`
method call.

The reason for this change is that computing roll, pitch, and yaw
involves several expensive trigonometric functions (sinf, cosf, asinf,
and two times atanf) not needed for the actual algorithm to
work. However, in many applications the update method is called at a
high rate (e.g. 500 hertz) but the actual estimates are only needed at
a much lower rate (e.g. 50 herz) by downstream systems. Hence, in this
example, 90% of those expensive trigonometric functions are
wasted.

Closes #7.